Unify Snippet- and ActionPicker into MultiPicker
Every Picker now has the same feature set, snippets can execute
a python callable *and* execute an action with the result of that
call. Every instance of a picker may choose a default behaviour
for simple section items via "simple-section-default" from one of:

 * value (which is the plain snippet feature)
 * callable (which is like a complex snippet with a callable option)
 * action (which behave like the action from the previous ActionPicker)

If, in complex sections, there is a callable *and* an action option,
both will be evaluated consecutively, passing the result on to the
next handler.

Additionally the interface was improved to allows adding more handlers
in the future. Although I cannot think of anything right now and the
first thing that I would want to implement will most definitly require
some changes to the interface. :-(

The old Snippet- and ActionPicker plugins have been kept as boiler-
plate, configuring a MultiPicker to mimic their previous behaviour
for convenience.