Split HistoryPicker for use it without a PopupPlugin
To be able to use the HistoryPicker core logic in other contexts,
we split off the history management logic into a HistoryController
class that supports, and is in fact based on the standard Plugin
interface. This allows for histories to be created at runtime from
within another plugin. There is currently no way to connect this
history and its signals to the outside world without connecting
the events manually in the instantiating plugin.

Additionally the PopupPlugin.popup method now accepts an items
parameter that, when provided, will be used instead of the items
method on the plugin. This allows for multiple popups from the same
plugin with different content without handcoding the logic of the
popup method.