Add a ClipboardExchange plugin to share your Clipboard
The plugin lets you configure shared hmac and aes secrets
to encrypt all traffic, as well as a list of users via
"user.<name> = ip[:port]". You can then suggest the contents
of your clipboard to one of those users via a popup menu.
The other user will then get a notification and can accept
the suggestion as his current clipboard contents.

Suggestions may be reviewed later on with another popup
menu that shows all suggestions on a per user basis.

The relevant options and defaults are configure in the
exmaple config file.