tray: Cleanup all tray icons and their menus
For this we picked an "official" icon from icons8.com and now refer to
it whenever we need our "logo", we still allow passing a custom tray
icon and still keep shipping the svgs for old times sake.

Since the icons8 license requires us to mention them in our about
dialog, we also add an about dialog to mention them on. :-D

The appindicator plugin needed a bit of tweaking since it also
needs porting to gtk3.